created separate navbar that is called in pages that need it. Added logout to navbar.
