+++ game/chip.css
@@ -111,9 +111,4 @@ html {
   margin: 0 20px;
 }
 
-.select_options {
-  background-color: gray;
-  padding: 20px;
-  margin: 10px;
-}
