+++ leaderboard/leaderboard_page.php
@@ -1,16 +1,17 @@
 <?php 
-require "../header.php";
+require "../database/header.php";
+include "../menu/navbar.html";
 ?>
 <!DOCTYPE html>
 <html>
     <head><link rel="stylesheet" href="../menu/menu.css"></head>
     <body>
-    <div class="topnav">
+    <!-- <div class="topnav">
         <a href="../connect4.php">Play</a>
         <a class="active" href="help.html">Instructions</a>
         <a href="../leaderboard/leaderboard_page.php">Leaderboard</a>
         <a href="contact_us.html">Contact Us</a>
-    </div> 
+    </div>  -->
         <form>
             <fieldset>
             <legend><h1>Leaderboard</h1></legend>
