+++ login/login_mysql.php
@@ -1,6 +1,6 @@
 <?php
 // connect to db
-require_once '../db_connect.php';
+require_once '../database/db_connect.php';
 
 session_start();
