+++ login/login_register_page.php
@@ -1,15 +1,11 @@
-
-<?php 
-
-session_start(); 
-?>
+<?php session_start(); ?>
 
 <!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <title>Connect4</title>
-    <link rel="stylesheet" href="../main_page.css" />
+    <link rel="stylesheet" href="main_page.css" />
 
   </head>
   <body>
