+++ menu/help.php
@@ -1,14 +1,16 @@
+
+<?php include "navbar.html"?>
 <!DOCTYPE html>
 <html>
 <head><link rel="stylesheet" href="menu.css"><head>
 
 <body>
-    <div class="topnav">
+    <!-- <div class="topnav">
         <a href="../connect4.php">Play</a>
         <a class="active" href="help.html">Instructions</a>
         <a href="../leaderboard/leaderboard_page.php">Leaderboard</a>
         <a href="contact_us.html">Contact Us</a>
-    </div> 
+    </div>  -->
 
     <h1>How To Play Connect 4</h1>
