+++ menu/menu.php
@@ -1,6 +1,6 @@
 
 <?php
-require "../header.php";
+require "../database/header.php";
 ?>
 
 <!DOCTYPE html>
@@ -8,21 +8,22 @@ require "../header.php";
 <head><link rel="stylesheet" href="menu.css"></head>
 
 <body>
+<h1>Welcome, <b><?php echo $_SESSION['username']; ?></b></h1>
     <section id="menu">
         <div id="navigate">
-            <a href="../game_options.php">Play Connect 4</a>
+            <a href="../game/game_options.php">Play Connect 4</a>
         </div>
         <div id="navigate">
-            <a href="help.html">Instructions</a>
+            <a href="../menu/help.php">Instructions</a>
         </div>
         <div id="navigate">
             <a href="../leaderboard/leaderboard_page.php">Leaderboard</a>
         </div>
         <div id="navigate">
-            <a href="contact_us.html">Contact Us</a>
+            <a href="../menu/contact_us.php">Contact Us</a>
         </div>
         <div id="navigate">
-            <a href="../login/login_register_page.php">Log Out</a>
+            <a href="../login/logout.php">Log Out</a>
         </div>
     </section>
 </body>
